+++ src/ni_measurement_plugin_converter/constants/_measui.py
@@ -2,13 +2,12 @@
 
 import ni_measurement_plugin_sdk_service as nims
 
-
 PIN_NAMES = "pin_names"
 SUPPORTED_NIMS_DATATYPES = [
     nims.DataType.Int64.name,
-    nims.DataType.Float.name,
+    nims.DataType.Double.name,
     nims.DataType.String.name,
     nims.DataType.Boolean.name,
     nims.DataType.Int64Array1D.name,
-    nims.DataType.FloatArray1D.name,
+    nims.DataType.DoubleArray1D.name,
 ]