+++ src/ni_measurement_plugin_converter/constants/_measurement_service.py
@@ -13,16 +13,17 @@ class DriverSession(Enum):
     niscope = "nims.session_management.INSTRUMENT_TYPE_NI_SCOPE"
     niswitch = "nims.session_management.INSTRUMENT_TYPE_NI_RELAY_DRIVER"
     nidaqmx = "nims.session_management.INSTRUMENT_TYPE_NI_DAQMX"
+    nivisa = "INSTRUMENT_TYPE_NI_VISA"
 
 
 # Python native data types and its corresponding `measurement_plugin_sdk_service` data types.
 NIMS_TYPE = {
     "int": "nims.DataType.Int64",
-    "float": "nims.DataType.Float",
+    "float": "nims.DataType.Double",
     "str": "nims.DataType.String",
     "bool": "nims.DataType.Boolean",
     "List[int]": "nims.DataType.Int64Array1D",
-    "List[float]": "nims.DataType.FloatArray1D",
+    "List[float]": "nims.DataType.DoubleArray1D",
     "List[str]": "nims.DataType.StringArray1D",
     "List[bool]": "nims.DataType.BooleanArray1D",
 }