+++ src/ni_measurement_plugin_converter/utils/_manage_session.py
@@ -16,6 +16,8 @@ from ni_measurement_plugin_converter.constants import (
 
 _RESERVATION = "reservation"
 _SESSION_INFO = "session_info"
+_SESSION_CONSTRUCTOR = "session_constructor"
+_INSTRUMENT_TYPE = "instrument_type"
 
 
 def manage_session(migrated_file_dir: str, function: str) -> str:
@@ -40,11 +42,15 @@ def manage_session(migrated_file_dir: str, function: str) -> str:
     code_tree = ast.parse(source_code)
 
     logger.info(UserMessage.ADD_RESERVE_SESSION)
-    reservation_added_source_code = add_reservation_param(code_tree, function)
+    reservation_added_code_tree = add_param(
+        code_tree=code_tree,
+        function=function,
+        param=_RESERVATION,
+    )
 
     logger.info(UserMessage.REPLACE_SESSION_INITIALIZATION)
-    session_replaced_source_code, session_details = replace_session_initialization(
-        source_code=reservation_added_source_code,
+    session_replaced_code_tree, session_details = replace_session_initialization(
+        code_tree=reservation_added_code_tree,
         function=function,
     )
 
@@ -53,55 +59,38 @@ def manage_session(migrated_file_dir: str, function: str) -> str:
         session_name = session_name
 
     logger.info(UserMessage.ASSIGN_SESSION_INFO)
-    session_inserted_source_code = insert_session_assignment(
-        source_code=session_replaced_source_code,
+
+    session_inserted_code_tree = insert_session_assignment(
+        code_tree=session_replaced_code_tree,
         function=function,
-        session_info=f"{session_name} = {_SESSION_INFO}.session",
+        session_name=session_name,
     )
 
+    code_tree = astor.to_source(session_inserted_code_tree)
+
     with open(migrated_file_dir, "w", encoding=ENCODING) as file:
-        file.write(session_inserted_source_code)
+        file.write(code_tree)
 
     logger.debug(DebugMessage.MIGRATED_FILE_MODIFIED)
 
     return instrument_type
 
 
-def add_reservation_param(code_tree: ast.Module, function: str) -> str:
-    """Add reservation parameter to source code.
+def replace_session_initialization(
+    code_tree: ast.Module,
+    function: str,
+) -> Tuple[str, List[Tuple[str, str, str]]]:
+    """Replace session initialization in the migrated file.
 
     Args:
         code_tree (ast.Module): Migrated measurement source code tree.
         function (str): Measurement function name.
 
     Returns:
-        str: Updated source code with reservation parameter.
-    """
-    for node in code_tree.body:
-        if isinstance(node, ast.FunctionDef) and node.name == function:
-            node.args.args.insert(0, _RESERVATION)
-
-    added_source = astor.to_source(code_tree)
-
-    return added_source
-
-
-def replace_session_initialization(
-    source_code: str,
-    function: str,
-) -> Tuple[str, List[Tuple[str, str, str]]]:
-    """Replace session initialization in the migrated file.
-
-    Args:
-        source_code (str): Migrated measurement source code.
-        function (str): Measurement function name.
-
-    Returns:
-        Tuple[str, List[Tuple[str, str]]]: Updated source code and List of tuple of \
+        Tuple[str, List[Tuple[str, str]]]: Updated source code tree and List of tuple of \
             replaced drivers, sessions.
     """
     replacements = []
-    code_tree = ast.parse(source_code)
 
     for node in ast.walk(code_tree):
         if isinstance(node, ast.FunctionDef) and node.name == function:
@@ -114,9 +103,38 @@ def replace_session_initialization(
                     ):
                         replacements.extend(__replace_session(child_node, driver.name))
 
-    modified_source = astor.to_source(code_tree)
+            if replacements and replacements[0][0] == DriverSession.nivisa.name:
+                instrument_type_added_code_tree = add_param(
+                    code_tree=code_tree,
+                    function=function,
+                    param=_INSTRUMENT_TYPE,
+                )
 
-    return modified_source, replacements
+                code_tree = add_param(
+                    code_tree=instrument_type_added_code_tree,
+                    function=function,
+                    param=_SESSION_CONSTRUCTOR,
+                )
+
+    return code_tree, replacements
+
+
+def add_param(code_tree: ast.Module, function: str, param: str) -> ast.Module:
+    """Add a parameter to user measurement function in code tree.
+
+    Args:
+        code_tree (ast.Module): Migrated measurement source code tree.
+        function (str): Measurement function name.
+        param (str): Parameter to be added.
+
+    Returns:
+        ast.Module: Updated measurement function with parameter.
+    """
+    for node in code_tree.body:
+        if isinstance(node, ast.FunctionDef) and node.name == function:
+            node.args.args.insert(0, param)
+
+    return code_tree
 
 
 def __replace_session(node: ast.With, driver: str) -> List[Tuple[str, str, str]]:
@@ -125,12 +143,7 @@ def __replace_session(node: ast.With, driver: str) -> List[Tuple[str, str, str]]
     for item in node.items:
         if isinstance(item.context_expr, ast.Call):
             call = item.context_expr
-            if (
-                isinstance(call.func, ast.Attribute)
-                and isinstance(call.func.value, ast.Name)
-                and call.func.value.id == driver
-                and call.func.attr == "Session"
-            ):
+            if __instrument_is_supported_ni_drivers(call, driver):
                 actual_session_name = item.optional_vars.id
 
                 item.optional_vars.id = _SESSION_INFO
@@ -147,12 +160,7 @@ def __replace_session(node: ast.With, driver: str) -> List[Tuple[str, str, str]]
                 call.keywords.clear()
                 call.args.clear()
 
-            elif (
-                isinstance(call.func, ast.Attribute)
-                and isinstance(call.func.value, ast.Name)
-                and call.func.value.id == driver
-                and call.func.attr == "Task"
-            ):
+            elif __instrument_is_ni_daqmx(call, driver):
                 actual_session_name = item.optional_vars.id
 
                 item.optional_vars.id = _SESSION_INFO
@@ -169,36 +177,98 @@ def __replace_session(node: ast.With, driver: str) -> List[Tuple[str, str, str]]
                 call.keywords.clear()
                 call.args.clear()
 
+            elif __instrument_is_visa_type(call):
+                actual_session_name = item.optional_vars.id
+
+                item.optional_vars.id = _SESSION_INFO
+                call.func.attr = "initialize_session"
+                call.func.value.id = _RESERVATION
+
+                replacements.append(
+                    (
+                        DriverSession.nivisa.name,
+                        actual_session_name,
+                    )
+                )
+
+                call.keywords.clear()
+                call.args.clear()
+
+                call.args = [
+                    ast.Name(id=_SESSION_CONSTRUCTOR, ctx=ast.Load()),
+                    ast.Name(id=_INSTRUMENT_TYPE, ctx=ast.Load()),
+                ]
+
     return replacements
 
 
+def __instrument_is_supported_ni_drivers(call: ast.Call, driver: str) -> bool:
+    if (
+        isinstance(call.func, ast.Attribute)
+        and isinstance(call.func.value, ast.Name)
+        and call.func.value.id == driver
+        and call.func.attr == "Session"
+    ):
+        return True
+
+    return False
+
+
+def __instrument_is_ni_daqmx(call: ast.Call, driver: str) -> bool:
+    if (
+        isinstance(call.func, ast.Attribute)
+        and isinstance(call.func.value, ast.Name)
+        and call.func.value.id == driver
+        and call.func.attr == "Task"
+    ):
+        return True
+
+    return False
+
+
+def __instrument_is_visa_type(call: ast.Call) -> bool:
+    if (
+        isinstance(call.func, ast.Attribute)
+        and isinstance(call.func.value, ast.Name)
+        and (
+            call.func.attr == "open_resource"
+            or call.func.attr == "get_instrument"
+            or call.func.attr == "instrument"
+        )
+    ):
+        return True
+
+    return False
+
+
 def insert_session_assignment(
-    source_code: str,
+    code_tree: ast.Module,
     function: str,
-    session_info: str,
-) -> None:
+    session_name: str,
+) -> ast.Module:
     """Insert session assignment into migrated measurement file source code.
 
     Args:
-        source_code (str): Migrated measurement source code.
+        source_code (ast.Module): Migrated measurement source code tree.
         function (str): Measurement function name.
-        session_info (str): Session information.
+        session_name (str): Session name from user measurement.
+
+    Returns:
+        ast.Module: Session assignment inserted source code tree.
     """
-    code_tree = ast.parse(source_code)
+    session_info = ast.Assign(
+        targets=[ast.Name(id=session_name, ctx=ast.Store())],
+        value=ast.Attribute(
+            value=ast.Name(id=_SESSION_INFO, ctx=ast.Load()),
+            attr="session",
+            ctx=ast.Load(),
+        ),
+    )
 
     for node in ast.walk(code_tree):
         if isinstance(node, ast.FunctionDef) and node.name == function:
-            for _, child_node in enumerate(node.body):
+            for child_node in node.body:
                 if isinstance(child_node, ast.With):
-                    # Get the start position of the `with` statement.
-                    indent = " " * (child_node.col_offset + 4)
+                    child_node.body.insert(0, session_info)
 
-                    # Construct the text to insert with proper indentation.
-                    text_line = f"{indent}{session_info}\n"
-
-                    # Insert the text immediately after the with block.
-                    source_lines = source_code.split("\n")
-                    source_lines.insert(child_node.lineno, text_line)
-                    modified_source = "\n".join(source_lines)
-
-    return modified_source
+    return code_tree